Use 64 bits when calculating an arena's handle.

+++ cf/include/arenax.h
@@ -146,7 +146,7 @@ static inline void
 cf_arenax_set_handle(cf_arenax_handle* h, uint32_t stage_id,
 		uint32_t element_id)
 {
-	*h = (stage_id << ELEMENT_ID_NUM_BITS) | element_id;
+	*h = ((uint64_t)stage_id << ELEMENT_ID_NUM_BITS) | element_id;
 }
 
 cf_arenax_err cf_arenax_add_stage(cf_arenax* _this);
